refactor: 카테고리 DiffUtil 분리
