+++ android/app/src/main/java/com/ddangddangddang/android/feature/register/category/MainCategoryAdapter.kt
@@ -2,16 +2,13 @@ package com.ddangddangddang.android.feature.register.category
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.ddangddangddang.android.databinding.ItemSelectMainCategoryBinding
 import com.ddangddangddang.android.model.CategoriesModel
 
 class MainCategoryAdapter(
     private val onItemClickListener: (Long) -> Unit,
-) : ListAdapter<CategoriesModel.CategoryModel, MainCategoryViewHolder>(
-    mainCategoryDiffUtil,
-) {
+) : ListAdapter<CategoriesModel.CategoryModel, MainCategoryViewHolder>(CategoryDiffUtil()) {
 
     fun setCategories(categories: List<CategoriesModel.CategoryModel>) {
         submitList(categories)
@@ -30,23 +27,5 @@ class MainCategoryAdapter(
 
     override fun onBindViewHolder(holder: MainCategoryViewHolder, position: Int) {
         holder.bind(currentList[position])
-    }
-
-    companion object {
-        private val mainCategoryDiffUtil = object : DiffUtil.ItemCallback<CategoriesModel.CategoryModel>() {
-            override fun areItemsTheSame(
-                oldItem: CategoriesModel.CategoryModel,
-                newItem: CategoriesModel.CategoryModel,
-            ): Boolean {
-                return oldItem.id == newItem.id
-            }
-
-            override fun areContentsTheSame(
-                oldItem: CategoriesModel.CategoryModel,
-                newItem: CategoriesModel.CategoryModel,
-            ): Boolean {
-                return oldItem == newItem
-            }
-        }
     }
 }
